Add:  1. delete user 2. login after registration

+++ src/com/raviv/coupons/dao/interfaces/IUsersDao.java
@@ -11,6 +11,6 @@ public interface IUsersDao {
 
 	public void 	updateUser	(User	user  ) throws ApplicationException;
 
-	public void 	deleteUser	(int	userId) throws ApplicationException;
+	public void 	deleteUser	(long	userId) throws ApplicationException;
 	
 }
